pyproject.toml
[project]
name = "project-nyx"
version = "0.1.0"
description = "NS Automation Framework"
authors = [
    {name = "niart120",email = "38847256+niart120@users.noreply.github.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pyserial >=3.5,<4.0",
    "pillow >=12.1.0,<13.0.0",
    "loguru >=0.7.3,<0.8.0",
    "cv2-enumerate-cameras >=1.1.18.3,<2.0.0.0",
    "tomlkit >=0.13.2,<0.14.0",
    "requests >=2.32.3,<3.0.0",
    "setuptools >=80.4.0",
    "opencv-python>=4.11.0.86",
    "pyside6>=6.9.0",
    "paddlepaddle>=3.0.0,<3.3.0",
    "paddleocr>=3.0.0",
    "chardet>=5.0.0,<6.0.0",
    "numpy>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-qt>=4.4.0",
    "ruff>=0.11.7",
]

[project.scripts]
nyx-cli = "nyxpy.cli.run_cli:main"
nyx-gui = "nyxpy.gui.run_gui:main"

[build-system]
requires = ["hatchling>=1.0"]
build-backend = "hatchling.build"

[tool.hatch.build]
targets.wheel.packages = ["src/nyxpy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "realdevice: tests requiring real hardware devices (capture card, serial port)",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade (flags from __future__ imports, old typing syntax, etc.)
]
ignore = [
    "E501",  # line too long — ruff format が担う
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E402"]  # pytest.importorskip / sys.path.insert 後の import を許容

